Move injected objects into their own package.
