+++ src/main/java/com/greenlaw110/di_benchmark/SplitStartupBenchmark.java
@@ -2,6 +2,7 @@ package com.greenlaw110.di_benchmark;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
+import com.greenlaw110.di_benchmark.objects.A;
 import dagger.ObjectGraph;
 import org.codejargon.feather.Feather;
 import org.osgl.inject.Genie;
