+++ src/main/java/com/greenlaw110/di_benchmark/objects/B.java
@@ -1,4 +1,4 @@
-package com.greenlaw110.di_benchmark;
+package com.greenlaw110.di_benchmark.objects;
 
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -8,30 +8,29 @@ import java.util.Objects;
 
 @Component
 @Scope("prototype")
-public class A {
-    private final B b;
+public class B {
+    private final C c;
 
     @Inject
-    public A(B b) {
-        this.b = b;
+    public B(C c) {
+        this.c = c;
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        A a = (A) o;
-        return Objects.equals(b, a.b);
+        B b = (B) o;
+        return Objects.equals(c, b.c);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(b);
+        return Objects.hash(c);
     }
 
     @Override
     public String toString() {
         return getClass().getSimpleName();
     }
-
 }
