+++ src/main/java/com/greenlaw110/di_benchmark/objects/D2.java
@@ -1,4 +1,4 @@
-package com.greenlaw110.di_benchmark;
+package com.greenlaw110.di_benchmark.objects;
 
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -8,31 +8,30 @@ import java.util.Objects;
 
 @Component
 @Scope("prototype")
-public class C {
-    private final D1 d1;
-    private final D2 d2;
+public class D2 {
+    private final E e;
 
     @Inject
-    public C(D1 d1, D2 d2) {
-        this.d1 = d1;
-        this.d2 = d2;
+    public D2(E e) {
+        this.e = e;
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        C c = (C) o;
-        return Objects.equals(d1, c.d1) &&
-                Objects.equals(d2, c.d2);
+        D2 d2 = (D2) o;
+        return Objects.equals(e, d2.e);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(d1, d2);
+        return Objects.hash(e);
     }
 
     @Override
     public String toString() {
         return getClass().getSimpleName();
-    }}
+    }
+
+}
